RF: Establish core GitRepo class
With the absolute bare essentials and pretty much no real functionality.
